fix offline mode not fully skipping npm packages
